+++ vusic/separation/scripts/__init__.py
@@ -1,1 +1,1 @@
-__all__ = ["preprocess", "stem_converter"]
+__all__ = ["preprocess", "stem_converter", "training"]
